refactor: added unit tests to user UseCase

+++ backend/src/UseCases/Student/DeleteStudent.ts
@@ -1,5 +1,5 @@
-import { StudentRepository } from "../../Infrastructure/repositories/StudentRepository";
-import { AppError } from "../../shared/errors/AppError";
+import {StudentRepository} from "../../Infrastructure/repositories/StudentRepository";
+import {AppError} from "../../shared/errors/AppError";
 
 export class DeleteStudent {
     private studentRepository: StudentRepository;
@@ -12,6 +12,10 @@ export class DeleteStudent {
         const student = await this.studentRepository.findById(id);
         if (!student) {
             throw new AppError("STUDENT_NOT_FOUND");
+        }
+
+        if (!Number.isInteger(id) || id <= 0) {
+            throw new AppError("INVALID_ID");
         }
 
         await this.studentRepository.delete(id);
